+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Added
 
 Changed
 ^^^^^^^
+- Update DataFrameMethodTransformer tests to have inheritable init class that can be used by othe test files.
 - Moved BaseTransformer, DataFrameMethodTransformer, BaseMappingTransformer, BaseMappingTransformerMixin and Mapping Transformer over to the new testing framework.
 - Refactored MappingTransformer by removing redundant init method.