+++ CHANGELOG.rst
@@ -37,6 +37,12 @@ Removed
 this functionality. This required updating a lot of test files.
 - The `columns_set_or_check()` method from BaseTransformer. With the above change it was no longer necessary. Subsequent updates to nominal transformers and their tests were required.
 
+1.2.3 (2024-02-23)
+------------------
+Added
+^^^^^
+- Split existing input check into two better defined checks for TwoColumnOperatorTransformer `#183 <https://github.com/lvgig/tubular/pull/183>`_
+- Created unit tests for checking column type and size `#183 <https://github.com/lvgig/tubular/pull/183>`_
 
 1.2.2 (2024-02-20)
 ------------------